Simplify and remove javascript from html_annotate templates.
This simplify and remove javascript from the annotate templates html,
and allow to expand individual lines instead of full functions.

This also only use details and summary elements, in order to make this
later compatible with html clients like JupyterLab and Nteract that do
not have global javascript context in which you can inject Javascript
without writing full extensions.

In particular we avoid to color tokens, but color background in
gree-ish/red-ish, this should allow to embed syntax highlighted code
and have proper coloring of tokens using pygments.

The style is a compromize betwen loking good as a standalone html file
and looking good when used in jupyterLab with the numba annotate magic.